Now Components and Actors have relative_location and Actor location in a center of sprite/collider

+++ source/engine/AI/poly2tri/sweep/sweep_context.h
@@ -62,7 +62,7 @@ Point* head();
 void set_tail(Point* p1);
 Point* tail();
 
-int point_count();
+size_t point_count();
 
 Node& LocateNode(Point& point);
 
@@ -152,7 +152,7 @@ inline AdvancingFront* SweepContext::front()
   return front_;
 }
 
-inline int SweepContext::point_count()
+inline size_t SweepContext::point_count()
 {
   return points_.size();
 }
